requirements.txt
flask
numpy
onnxruntime
opencv-python
pillow
